backend/pyproject.toml
[project]
name = "heypi"
version = "0.1.0"
description = "An API."
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "argon2-cffi>=23.1.0",
  "asyncpg>=0.30.0",
  "fastapi>=0.115.3",
  "granian>=1.6.1",
  "pydantic[email]>=2.9.2",
  "pydantic-settings>=2.6",
  "uvloop>=0.21",
]

[tool.uv]
dev-dependencies = ["granian[reload]>=1.6.1"]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["COM812", "ISC001", "D104"]

pydocstyle.convention = "google"
